@@ -12,27 +12,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "14.0.0-rc.2",
-    "@angular/cdk": "^14.0.0-rc.1",
-    "@angular/common": "14.0.0-rc.2",
-    "@angular/compiler": "14.0.0-rc.2",
-    "@angular/core": "14.0.0-rc.2",
-    "@angular/forms": "14.0.0-rc.2",
-    "@angular/material": "^14.0.0-rc.1",
-    "@angular/platform-browser": "14.0.0-rc.2",
-    "@angular/platform-browser-dynamic": "14.0.0-rc.2",
-    "@angular/router": "14.0.0-rc.2",
+    "@angular/animations": "15.0.1",
+    "@angular/cdk": "^15.0.0",
+    "@angular/common": "15.0.1",
+    "@angular/compiler": "15.0.1",
+    "@angular/core": "15.0.1",
+    "@angular/forms": "15.0.1",
+    "@angular/material": "^15.0.0",
+    "@angular/platform-browser": "15.0.1",
+    "@angular/platform-browser-dynamic": "15.0.1",
+    "@angular/router": "15.0.1",
     "core-js": "^2.6.1",
-    "ngx-mat-select-search": "^4.2.0",
+    "ngx-mat-select-search": "^7.0.0",
     "rxjs": "6.5.4",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~14.2.9",
-    "@angular/cli": "^14.0.0-next.0",
-    "@angular/compiler-cli": "14.0.0-rc.2",
-    "@angular/language-service": "14.0.0-rc.2",
+    "@angular-devkit/build-angular": "~15.0.1",
+    "@angular/cli": "^15.0.1",
+    "@angular/compiler-cli": "15.0.1",
+    "@angular/language-service": "15.0.1",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.6",
     "@types/node": "^12.11.1",
@@ -40,6 +40,6 @@
     "jasmine-spec-reporter": "~5.0.0",
     "ts-node": "~7.0.1",
     "tslint": "~6.1.0",
-    "typescript": "4.7.2"
+    "typescript": "4.8.4"
   }
 }
